+++ SSLTest/src/scan_vulnerabilities/utils.py
